refactor : [미션번외] FSD구조로 리팩토링
