+++ src/features/themeCatalog/index.ts
@@ -1,2 +1,3 @@
 export { default as ThemeField } from './ui/ThemeField';
-export * from './model/useThemeProducts';
+export { default as ThemeHeroSection } from './ui/ThemeHeroSection';
+export { default as ThemeProductSection } from './ui/ThemeProductSection';
