+++ src/features/themeCatalog/ui/ThemeHeroSection/index.tsx
