+++ src/features/themeCatalog/ui/ThemeHeroSection/styles.ts
