+++ src/widgets/index.ts
@@ -1,5 +1,3 @@
 export * from './navigation';
-export * from './themeHero';
-export * from './ThemeCatalog';
-export * from './themeProductList';
+export * from './themeSection';
 export * from './orderForm';
