--- src/widgets/themeHero/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ThemeHeroSection } from './ThemeHeroSection';
