--- src/widgets/themeProductList/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ThemeProductSection } from './ThemeProductSection';
