+++ src/widgets/themeSection/ui/CheerUpMessage/index.tsx
