+++ src/widgets/themeSection/ui/CheerUpMessage/styles.ts
