+++ src/widgets/themeSection/ui/ThemeSection/index.tsx
