+++ src/widgets/themeSection/ui/ThemeSection/styles.ts
